allergy-inspector-main/utils/media_handler.py
import base64

import streamlit as st
from PIL import Image
import io



def image_to_base64(image_bytes):
    """
    Converts an image to a base64-encoded string.

    Args:
        image_path (str): The path to the image file.

    Returns:
        str: The base64-encoded string of the image.
    """
    try:
        base64_string = base64.b64encode(image_bytes).decode("utf-8")
        return base64_string
    except FileNotFoundError:
        return "Image file not found. Please check the path."
    except Exception as e:
        return f"An error occurred: {str(e)}"