allergy-inspector-main/requirements.txt
streamlit
openai
streamlit_chat
together
python-dotenv
